Move omf.activity.helpers -> omf.helpers.
I think this test hierarchy is mistakenly in omf.activity, given that none of
them specifically test activities.
