+++ app/src/androidTest/java/org/mozilla/focus/activity/PullDownToRefreshTest.java
@@ -19,7 +19,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mozilla.focus.R;
-import org.mozilla.focus.activity.helpers.MainActivityFirstrunTestRule;
+import org.mozilla.focus.helpers.MainActivityFirstrunTestRule;
 
 import java.io.IOException;
