+++ app/src/androidTest/java/org/mozilla/focus/activity/SearchEngineSelectionTest.java
@@ -25,7 +25,7 @@ import static android.support.test.espresso.action.ViewActions.click;
 import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertTrue;
 import static org.mozilla.focus.activity.TestHelper.waitingTime;
-import static org.mozilla.focus.activity.helpers.EspressoHelper.openSettings;
+import static org.mozilla.focus.helpers.EspressoHelper.openSettings;
 import static org.mozilla.focus.fragment.FirstrunFragment.FIRSTRUN_PREF;
 
 // This test checks the search engine can be changed
