+++ app/src/androidTest/java/org/mozilla/focus/activity/SettingsAppearanceTest.java
@@ -21,7 +21,7 @@ import org.junit.runner.RunWith;
 
 import static junit.framework.Assert.assertTrue;
 import static org.mozilla.focus.activity.TestHelper.waitingTime;
-import static org.mozilla.focus.activity.helpers.EspressoHelper.openSettings;
+import static org.mozilla.focus.helpers.EspressoHelper.openSettings;
 import static org.mozilla.focus.fragment.FirstrunFragment.FIRSTRUN_PREF;
 
 // This test checks all the headings in the Settings menu are there
