+++ app/src/androidTest/java/org/mozilla/focus/activity/SwitchLocaleTest.java
@@ -30,8 +30,8 @@ import java.util.Locale;
 import static android.os.Build.VERSION.SDK_INT;
 import static android.support.test.espresso.action.ViewActions.click;
 import static org.mozilla.focus.activity.TestHelper.waitingTime;
-import static org.mozilla.focus.activity.helpers.EspressoHelper.openMenu;
-import static org.mozilla.focus.activity.helpers.EspressoHelper.openSettings;
+import static org.mozilla.focus.helpers.EspressoHelper.openMenu;
+import static org.mozilla.focus.helpers.EspressoHelper.openSettings;
 import static org.mozilla.focus.fragment.FirstrunFragment.FIRSTRUN_PREF;
 
 // This test checks all the headings in the Settings menu are there
