+++ app/src/androidTest/java/org/mozilla/focus/activity/ToggleBlockTest.java
@@ -19,7 +19,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mozilla.focus.R;
-import org.mozilla.focus.activity.helpers.SessionLoadedIdlingResource;
+import org.mozilla.focus.helpers.SessionLoadedIdlingResource;
 
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
