+++ app/src/androidTest/java/org/mozilla/focus/activity/screenshots/BrowserScreenScreenshots.java
@@ -23,7 +23,7 @@ import org.junit.runner.RunWith;
 import org.mozilla.focus.R;
 import org.mozilla.focus.activity.MainActivity;
 import org.mozilla.focus.activity.TestHelper;
-import org.mozilla.focus.activity.helpers.MainActivityFirstrunTestRule;
+import org.mozilla.focus.helpers.MainActivityFirstrunTestRule;
 
 import java.io.IOException;
